WIP:
GameMethods Class
Changed method getReverseLocationOfY to accept int instead of Point.
Changed method replaceObjects to get the shortest path where there are no blockages to lead the objects to thier final destination.
Created method addNewObjects to retrieve the shortest path.
Created method blockedPositions which returns a list of blocked elements that are not empty.

PositionList Class
Additional functions added to get or input an element to the Point ArrayList at an identified index